move sympathy to component
